Maps demo
This code demonstrates animation behaviors involving MapKit. The code alternates between two different, nearby points of interest, replicating the behavior of having an offset on rotation, as seen in both the Apple Maps and Google Maps apps. The animation moving to the new POI can occur simultaneously, representing an "idealized" UX experience, or subsequently after the rotation, replicating the effect of Google Maps.